Move the scan-row assembly out of AppListModel
The rows a disk scan produces — the unchecked ones and the merge of a fresh
scan onto what is already on screen — are the code in this repo that has been
got wrong most often. Four of the thirteen defects found reviewing the channel
store lived here, three of them introduced while fixing the previous one, each
compiling and each passing make test, because nothing executes App/Sources.

ScanRowAssembly is a plain enum of static functions with no UI, no Preferences
and no AppListModel, so a test target can compile this one file rather than
the whole app. Channel proofs arrive through a ChannelProofSource protocol:
ResolvedChannelStore.Snapshot drops every entry whose path is not on disk, so
testing against the real type would mean minting bundles in the filesystem to
exercise an in-memory decision.

Fixes a fifth instance of the same defect while collapsing the call sites: the
merge built a bare UpdateResult for an app with no prior row — losing the
proven channel for a copy installed while DuoUpdater is running, or a second
copy a previous scan de-duplicated away — with the proofs in scope three lines
below. Every unchecked row now has one construction path, so the omission has
nowhere left to happen. Not shown to be reproducible in the product; it is the
same shape as the four that were.
